server/requirements.txt
Flask~=2.2.2
bcrypt~=4.0.0
PyJWT~=2.4.0
psycopg[binary]~=3.0.16
gunicorn~=20.1.0